requirements.txt
customtkinter==5.2.2
tkinterdnd2==0.4.3
sounddevice==0.5.5
Pillow>=10.0.0
pyinstaller>=6.19.0
